docker-service/python_stock_prediction/python_requirements.txt
scikit-learn
joblib
torch
confluent-kafka
cassandra-driver
pandas
numpy